simple move go file ouside and use cmd/ folder
