+++ assets/metadata/trending/trending_videos_metadata.csv
@@ -1,0 +1,86 @@
+,video_id,video_url,trending_position,trending_date,trending_datetime
+0,4ZX9T0kWb4Y,https://youtu.be/4ZX9T0kWb4Y,1,2023-04-04,2023-04-04 21:52:11
+1,vS3_72Gb-bI,https://youtu.be/vS3_72Gb-bI,2,2023-04-04,2023-04-04 21:52:11
+2,Tp_YZNqNBhw,https://youtu.be/Tp_YZNqNBhw,3,2023-04-04,2023-04-04 21:52:11
+3,bCoV7VaobfA,https://youtu.be/bCoV7VaobfA,4,2023-04-04,2023-04-04 21:52:11
+4,-buJzSym2f8,https://youtu.be/-buJzSym2f8,5,2023-04-04,2023-04-04 21:52:11
+5,9XVBTaZ9F6Y,https://youtu.be/9XVBTaZ9F6Y,6,2023-04-04,2023-04-04 21:52:11
+6,3zUPDkt8aEI,https://youtu.be/3zUPDkt8aEI,7,2023-04-04,2023-04-04 21:52:11
+7,xS5e2o6REb4,https://youtu.be/xS5e2o6REb4,8,2023-04-04,2023-04-04 21:52:11
+8,mO0OuR26IZM,https://youtu.be/mO0OuR26IZM,9,2023-04-04,2023-04-04 21:52:11
+9,N4lnt6Oudq0,https://youtu.be/N4lnt6Oudq0,10,2023-04-04,2023-04-04 21:52:11
+10,Xq1lzZAEEPw,https://youtu.be/Xq1lzZAEEPw,11,2023-04-04,2023-04-04 21:52:11
+11,yPnGZBpvjN0,https://youtu.be/yPnGZBpvjN0,12,2023-04-04,2023-04-04 21:52:11
+12,GOIvv9Qbbp4,https://youtu.be/GOIvv9Qbbp4,13,2023-04-04,2023-04-04 21:52:11
+13,lUSpuGgZTTA,https://youtu.be/lUSpuGgZTTA,14,2023-04-04,2023-04-04 21:52:11
+14,IYMh0ddLrQM,https://youtu.be/IYMh0ddLrQM,15,2023-04-04,2023-04-04 21:52:11
+15,1WEAJ-DFkHE,https://youtu.be/1WEAJ-DFkHE,16,2023-04-04,2023-04-04 21:52:11
+16,XIOoqJyx8E4,https://youtu.be/XIOoqJyx8E4,17,2023-04-04,2023-04-04 21:52:11
+17,l4w6808wJcU,https://youtu.be/l4w6808wJcU,18,2023-04-04,2023-04-04 21:52:11
+18,7Z-owxKt3w0,https://youtu.be/7Z-owxKt3w0,19,2023-04-04,2023-04-04 21:52:11
+19,Yvqz-BYBnp4,https://youtu.be/Yvqz-BYBnp4,20,2023-04-04,2023-04-04 21:52:11
+20,Uy8vawDSg3g,https://youtu.be/Uy8vawDSg3g,21,2023-04-04,2023-04-04 21:52:11
+21,hmzNIiJ0jnU,https://youtu.be/hmzNIiJ0jnU,22,2023-04-04,2023-04-04 21:52:11
+22,cHXB-5woeHw,https://youtu.be/cHXB-5woeHw,23,2023-04-04,2023-04-04 21:52:11
+23,FExS3lFEwqc,https://youtu.be/FExS3lFEwqc,24,2023-04-04,2023-04-04 21:52:11
+24,7QNKS2wUGkQ,https://youtu.be/7QNKS2wUGkQ,25,2023-04-04,2023-04-04 21:52:11
+25,vkul44BkVMU,https://youtu.be/vkul44BkVMU,26,2023-04-04,2023-04-04 21:52:11
+26,5cJK9VJ24hQ,https://youtu.be/5cJK9VJ24hQ,27,2023-04-04,2023-04-04 21:52:11
+27,G6FARknIG6w,https://youtu.be/G6FARknIG6w,28,2023-04-04,2023-04-04 21:52:11
+28,Csi35H8_oDo,https://youtu.be/Csi35H8_oDo,29,2023-04-04,2023-04-04 21:52:11
+29,Fpn1imb9qZg,https://youtu.be/Fpn1imb9qZg,30,2023-04-04,2023-04-04 21:52:11
+30,bMRVJ64lzw0,https://youtu.be/bMRVJ64lzw0,31,2023-04-04,2023-04-04 21:52:11
+31,n0LPccQzr04,https://youtu.be/n0LPccQzr04,32,2023-04-04,2023-04-04 21:52:11
+32,T3xRxzCS6vw,https://youtu.be/T3xRxzCS6vw,33,2023-04-04,2023-04-04 21:52:11
+33,Vck5O1IXSWw,https://youtu.be/Vck5O1IXSWw,34,2023-04-04,2023-04-04 21:52:11
+34,_BZIxt07UNU,https://youtu.be/_BZIxt07UNU,35,2023-04-04,2023-04-04 21:52:11
+35,zrmLaMBJsdg,https://youtu.be/zrmLaMBJsdg,36,2023-04-04,2023-04-04 21:52:11
+36,Xoh9RY9uLdk,https://youtu.be/Xoh9RY9uLdk,37,2023-04-04,2023-04-04 21:52:11
+37,Ve9Swo-Zpek,https://youtu.be/Ve9Swo-Zpek,38,2023-04-04,2023-04-04 21:52:11
+38,WjDa3BJO3V4,https://youtu.be/WjDa3BJO3V4,39,2023-04-04,2023-04-04 21:52:11
+39,wwE1fZmIkss,https://youtu.be/wwE1fZmIkss,40,2023-04-04,2023-04-04 21:52:11
+40,YfZhHYrI1ZM,https://youtu.be/YfZhHYrI1ZM,41,2023-04-04,2023-04-04 21:52:11
+41,drLZxyv79Oo,https://youtu.be/drLZxyv79Oo,42,2023-04-04,2023-04-04 21:52:11
+42,2G84mU3WPaE,https://youtu.be/2G84mU3WPaE,43,2023-04-04,2023-04-04 21:52:11
+43,epgZ2Dbr_Tw,https://youtu.be/epgZ2Dbr_Tw,44,2023-04-04,2023-04-04 21:52:11
+44,NGZaVh3LqYo,https://youtu.be/NGZaVh3LqYo,45,2023-04-04,2023-04-04 21:52:11
+45,DpfNilof0Ag,https://youtu.be/DpfNilof0Ag,46,2023-04-04,2023-04-04 21:52:11
+46,HALG3Ty52mY,https://youtu.be/HALG3Ty52mY,47,2023-04-04,2023-04-04 21:52:11
+47,RK93dlRjJRY,https://youtu.be/RK93dlRjJRY,48,2023-04-04,2023-04-04 21:52:11
+48,sMgvKDIQ5YQ,https://youtu.be/sMgvKDIQ5YQ,49,2023-04-04,2023-04-04 21:52:11
+49,4gaHphhEQOs,https://youtu.be/4gaHphhEQOs,50,2023-04-04,2023-04-04 21:52:11
+50,brsmungVdYc,https://youtu.be/brsmungVdYc,51,2023-04-04,2023-04-04 21:52:11
+51,YudHcBIxlYw,https://youtu.be/YudHcBIxlYw,52,2023-04-04,2023-04-04 21:52:11
+52,9iOJjo1vU1A,https://youtu.be/9iOJjo1vU1A,54,2023-04-04,2023-04-04 21:52:11
+53,_bOPd9oHrGg,https://youtu.be/_bOPd9oHrGg,55,2023-04-04,2023-04-04 21:52:11
+54,RPH2APBoVtI,https://youtu.be/RPH2APBoVtI,57,2023-04-04,2023-04-04 21:52:11
+55,oC5Z-IJwwak,https://youtu.be/oC5Z-IJwwak,58,2023-04-04,2023-04-04 21:52:11
+56,Prbszu3u42g,https://youtu.be/Prbszu3u42g,59,2023-04-04,2023-04-04 21:52:11
+57,7vEPbxsnpyM,https://youtu.be/7vEPbxsnpyM,61,2023-04-04,2023-04-04 21:52:11
+58,AjTFhTgfecQ,https://youtu.be/AjTFhTgfecQ,62,2023-04-04,2023-04-04 21:52:11
+59,LSIOcCcEVaE,https://youtu.be/LSIOcCcEVaE,64,2023-04-04,2023-04-04 21:52:11
+60,rJNqj_QNkWM,https://youtu.be/rJNqj_QNkWM,65,2023-04-04,2023-04-04 21:52:11
+61,ZWliqIonBcw,https://youtu.be/ZWliqIonBcw,67,2023-04-04,2023-04-04 21:52:11
+62,Ey_vGhau7ho,https://youtu.be/Ey_vGhau7ho,69,2023-04-04,2023-04-04 21:52:11
+63,x0EauOLqR2g,https://youtu.be/x0EauOLqR2g,70,2023-04-04,2023-04-04 21:52:11
+64,rtC1vV1-MoY,https://youtu.be/rtC1vV1-MoY,71,2023-04-04,2023-04-04 21:52:11
+65,RbvFZX1-H_k,https://youtu.be/RbvFZX1-H_k,72,2023-04-04,2023-04-04 21:52:11
+66,p4jmzZG_zjY,https://youtu.be/p4jmzZG_zjY,74,2023-04-04,2023-04-04 21:52:11
+67,CuRXm6q0cLw,https://youtu.be/CuRXm6q0cLw,75,2023-04-04,2023-04-04 21:52:11
+68,fjSN8ig_EHI,https://youtu.be/fjSN8ig_EHI,76,2023-04-04,2023-04-04 21:52:11
+69,aKVvdgfbmWw,https://youtu.be/aKVvdgfbmWw,77,2023-04-04,2023-04-04 21:52:11
+70,CMB109kZuGM,https://youtu.be/CMB109kZuGM,78,2023-04-04,2023-04-04 21:52:11
+71,u6wOyMUs74I,https://youtu.be/u6wOyMUs74I,80,2023-04-04,2023-04-04 21:52:11
+72,nY3EpKcWnSY,https://youtu.be/nY3EpKcWnSY,81,2023-04-04,2023-04-04 21:52:11
+73,9oSm0c6bPTM,https://youtu.be/9oSm0c6bPTM,82,2023-04-04,2023-04-04 21:52:11
+74,a6qna-ZCbxA,https://youtu.be/a6qna-ZCbxA,83,2023-04-04,2023-04-04 21:52:11
+75,ZLLcxNiQ9Yg,https://youtu.be/ZLLcxNiQ9Yg,84,2023-04-04,2023-04-04 21:52:11
+76,Fv_BwXW7TTU,https://youtu.be/Fv_BwXW7TTU,85,2023-04-04,2023-04-04 21:52:11
+77,961v0E3b01g,https://youtu.be/961v0E3b01g,87,2023-04-04,2023-04-04 21:52:11
+78,XVRlNJxakqA,https://youtu.be/XVRlNJxakqA,88,2023-04-04,2023-04-04 21:52:11
+79,KCHT6a53OqY,https://youtu.be/KCHT6a53OqY,89,2023-04-04,2023-04-04 21:52:11
+80,ZnbXqntTWcw,https://youtu.be/ZnbXqntTWcw,91,2023-04-04,2023-04-04 21:52:11
+81,27FswS3KESk,https://youtu.be/27FswS3KESk,93,2023-04-04,2023-04-04 21:52:11
+82,qGJLTDxnMs8,https://youtu.be/qGJLTDxnMs8,95,2023-04-04,2023-04-04 21:52:11
+83,YLt73w6criQ,https://youtu.be/YLt73w6criQ,96,2023-04-04,2023-04-04 21:52:11
+84,buKcergt4dA,https://youtu.be/buKcergt4dA,97,2023-04-04,2023-04-04 21:52:11
