--- src/tube/go.mod
@@ -1,10 +1,0 @@
-module github.com/gpsyrou/tube-virality
-
-go 1.19
-
-require github.com/PuerkitoBio/goquery v1.8.1
-
-require (
-	github.com/andybalholm/cascadia v1.3.1 // indirect
-	golang.org/x/net v0.7.0 // indirect
-)
